Jent ES: use buffer that is filled with a thread
The Jitter RNG may be a bit too slow for some use cases. Therefore, use a
buffer that is filled by a separate Jitter RNG ES thread. Callers are
provided with data from the buffer and thus do not incur the wait time.

Yet, if the buffer is empty or the buffer logic is disabled, use the
Jitter RNG directly. Per default, the buffer handling is disabled at the
moment as the code is not fully and thoroughly tested. It can be enabled
by setting the variable esdm_jent_entropy_buffer_enable to true. Eventually
this variable will become a start-time option.

Signed-off-by: Stephan Mueller <smueller@chronox.de>